Revise keychain implementation
