+++ Sources/Scyther/User Interface/Menu/MenuViewModel.swift
@@ -177,10 +177,10 @@ internal class MenuViewModel {
         var securitySection: Section = Section()
         securitySection.title = "Security"
         securitySection.rows.append(actionRow(name: "Keychain Browser",
-                                                 icon: UIImage(systemImage: "key"),
-                                                 actionBlock: { [weak self] in
-                                                     self?.delegate?.viewModel(viewModel: self, shouldShowViewController: KeychainBrowserViewController())
-                                                 }))
+                                              icon: UIImage(systemImage: "key"),
+                                              actionBlock: { [weak self] in
+                                                  self?.delegate?.viewModel(viewModel: self, shouldShowViewController: DataBrowserViewController(data: KeychainBrowser.keychainItems))
+                                              }))
 
         /// Setup Support Section
         var supportSection: Section = Section()
